Add HW to a HSM group now working and creating group as well.

+++ src/cli/commands/add_hw_component_cluster.rs
@@ -15,10 +15,33 @@ pub async fn exec(
     target_hsm_group_name: &str,
     parent_hsm_group_name: &str,
     pattern: &str,
-    dryrun: bool,
+    nodryrun: bool,
     create_hsm_group: bool
 ) {
     let pattern = format!("{}:{}", target_hsm_group_name, pattern);
+
+    match mesa::hsm::group::mesa::http_client::get(shasta_token,
+                                                   shasta_base_url,
+                                                   shasta_root_cert,
+                                                   Some(&target_hsm_group_name.to_string())).await {
+        Ok(_) => log::debug!("The HSM group {} exists, good.",target_hsm_group_name),
+        Err(_) => {
+            if create_hsm_group {
+                log::info!("HSM group {} does not exist, but the option to create the group has been selected, creating it now.", target_hsm_group_name.to_string());
+                if nodryrun {
+                    mesa::hsm::group::mesa::http_client::create_new_hsm_group(shasta_token, shasta_base_url, shasta_root_cert, target_hsm_group_name, &[], "false", "", &[])
+                        .await.expect("Unable to create new HSM group");
+                } else {
+                    log::error!("Dryrun selected, cannot create the new group and continue.");
+                    std::process::exit(1);
+                }
+            }
+            else {
+                log::error!("HSM group {} does not exist, but the option to create the group was NOT specificied, cannot continue.", target_hsm_group_name.to_string());
+                std::process::exit(1);
+            }
+        }
+    };
 
     log::info!("pattern: {}", pattern);
 
@@ -238,12 +261,15 @@ pub async fn exec(
 
     // *********************************************************************************************************
     // UPDATE HSM GROUP MEMBERS IN CSM
-    for xname in nodes_moved_from_parent_hsm {
-        let _ = hsm::group::shasta::http_client::delete_member(shasta_token, shasta_base_url, shasta_root_cert, parent_hsm_group_name, &xname).await;
+    if !nodryrun {
+        log::info!("Dryrun enabled, not modifying the HSM groups on the system.")
+    } else {
+        for xname in nodes_moved_from_parent_hsm {
+            let _ = hsm::group::shasta::http_client::delete_member(shasta_token, shasta_base_url, shasta_root_cert, parent_hsm_group_name, &xname).await;
 
-        let _ = hsm::group::shasta::http_client::post_member(shasta_token, shasta_base_url, shasta_root_cert, target_hsm_group_name, &xname).await;
+            let _ = hsm::group::shasta::http_client::post_member(shasta_token, shasta_base_url, shasta_root_cert, target_hsm_group_name, &xname).await;
+        }
     }
-
     let target_hsm_group_value = serde_json::json!({
         "label": target_hsm_group_name,
         "decription": "",
